FEATURE/TL19-RECORD_CREATE_FULL: feature(api): refacror save or update record
### Summary
- **Removal of Deprecated DTO Classes**: `ServiceInfoDto` and related classes were removed.
- **Changes in Controller**: `RecordController` now uses `OptionalRecordCreateOrUpdateRequest`.
- **Service Logic Changes**: `RecordService` methods have been updated to use new entities.
- **Mapper Adjustments**: Mappers such as `RecordMapper` and `ServiceInfoMapper` were modified, removing the old mappings related to deprecated classes.

### Detailed Analysis

#### Controller Layer
- **RecordController**:
  - The controller now accepts `OptionalRecordCreateOrUpdateRequest`.
  - Method call from `service.createRecord(...)` was changed to `recordService.saveOrUpdate(...)`.

#### Service Layer
- **RecordService**:
  - Methods like `createRecord` and related ones seem to have been replaced with a more generalized method: `saveOrUpdate`.

#### Entity DTOs
- **Deprecated Classes Removed**:
  - `ServiceInfoDto` is no longer used.
  - `RecordCreateOrUpdateRequest` was replaced by `OptionalRecordCreateOrUpdateRequest`.
- This seems like part of an effort to modernize the API or make it more flexible with optional fields.

#### Mappers
- **Changes in Mapping**:
  - `RecordMapper` removed mappings for deprecated classes.
  - Similarly, `ServiceInfoMapper` cleaned up old mapping references.

### Test Layer
- **JooqAbacServiceTest**:
  - The test methods now call `.block()` on the results of ABAC checks to handle reactive programming style (likely using Project Reactor).

### Impact and Considerations

1. **Backward Compatibility**: Any clients or consumers relying on the old `RecordCreateOrUpdateRequest` will break unless they are updated as well.

2. **Dependency Updates**: Ensure that any dependent services, microservices, or external systems using these endpoints are also updated to reflect these changes.

3. **Testing**:
   - All unit and integration tests should be reviewed and possibly rewritten to account for these new methods and data structures.
   - The reactive programming style in the test layer might suggest a broader shift in architecture towards non-blocking I/O, which needs thorough testing.

4. **Documentation**: Update any API documentation, OpenAPI specs, or Swagger documentation to reflect the changes.

+++ src/main/java/com/tlback/web/dto/records/RecordPreviewResponse.java
@@ -9,5 +9,5 @@ public class RecordPreviewResponse {
     private Long id;
     private LocalDateTime tsFrom;
     private LocalDateTime tsTo;
-    private Integer serviceInfoId;
+    private Long serviceInfoId;
 }
